update code that only superusers to have access

+++ products/views.py
@@ -106,6 +106,11 @@ def product_detail(request, product_id):
 # render an empty instance of form so can be seen how it looks
 def add_product(request):
     """ Add a product to the store """
+    # only superusers to have access to the add, edit, and delete views
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     # the post handler
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -131,6 +136,11 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    # only superusers to have access to the add, edit, and delete views
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -156,6 +166,11 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    # only superusers to have access to the add, edit, and delete views
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
